Build compiles on client

+++ src/runtime_src/xdp/profile/database/static_info/aie_constructs.h
@@ -86,17 +86,17 @@ enum class module_type {
     // Loginal name
     std::string logicalName;
     // Column where I/O is mapped
-    short shimColumn;
+    uint8_t shimColumn;
     // slave or master - 0:slave, 1:master
-    short slaveOrMaster;
+    uint8_t slaveOrMaster;
     // Shim stream switch port id
-    short streamId;
+    uint8_t streamId;
     // Channel number
-    short channelNum;
+    uint8_t channelNum;
     // Burst length
-    short burstLength;
+    uint8_t burstLength;
     // I/O type - 0:PLIO, 1:GMIO
-    short type;
+    uint8_t type;
   };  
 
   /*
@@ -138,13 +138,13 @@ enum class module_type {
   struct TraceGMIO
   {
     uint32_t id;
-    uint16_t shimColumn;
-    uint16_t channelNumber;
-    uint16_t streamId;
-    uint16_t burstLength;
+    uint8_t shimColumn;
+    uint8_t channelNumber;
+    uint8_t streamId;
+    uint8_t burstLength;
 
-    TraceGMIO(uint32_t i, uint16_t col, uint16_t num, 
-              uint16_t stream, uint16_t len)
+    TraceGMIO(uint32_t i, uint8_t col, uint8_t num, 
+              uint8_t stream, uint8_t len)
       : id(i)
       , shimColumn(col)
       , channelNumber(num)
