+++ src/runtime_src/xdp/profile/database/static_info/device_info.h
@@ -131,8 +131,8 @@ namespace xdp {
 
     // ****** Functions for AIE information on the current xclbin ******
     XDP_CORE_EXPORT
-    void addTraceGMIO(uint32_t i, uint16_t col, uint16_t num, uint16_t stream,
-                      uint16_t len) ;
+    void addTraceGMIO(uint32_t i, uint8_t col, uint8_t num, uint8_t stream,
+                      uint8_t len) ;
     XDP_CORE_EXPORT
     void addAIECounter(uint32_t i, uint16_t col, uint16_t r, uint8_t num,
                        uint16_t start, uint16_t end, uint8_t reset,
